renamed the package for the current implementation encoding approach
