--- opendse-encoding/src/main/java/net/sf/opendse/encoding/constraints/SpecificationConstraintsMulti.java
@@ -1,75 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2015 OpenDSE
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- *******************************************************************************/
-package net.sf.opendse.encoding.constraints;
-
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.Set;
-
-import net.sf.opendse.model.Specification;
-import net.sf.opendse.model.parameter.ParameterReference;
-
-import org.opt4j.satdecoding.Constraint;
-import org.opt4j.satdecoding.Model;
-
-import com.google.inject.Inject;
-
-// TODO: Construct with "Specification"
-public class SpecificationConstraintsMulti implements SpecificationConstraints {
-
-	protected final Set<SpecificationConstraints> specificationConstraints;
-	
-	@Inject
-	public SpecificationConstraintsMulti(Set<SpecificationConstraints> specificationConstraints) {
-		super();
-		this.specificationConstraints = specificationConstraints;
-	}
-
-	@Override
-	public void doEncoding(Collection<Constraint> constraints) {
-		for(SpecificationConstraints sc: specificationConstraints){
-			sc.doEncoding(constraints);
-		}
-	}
-
-	@Override
-	public void doInterpreting(Specification implementation, Model model) {
-		for(SpecificationConstraints sc: specificationConstraints){
-			sc.doInterpreting(implementation, model);
-		}
-	}
-
-	@Override
-	public Set<ParameterReference> getActiveParameters() {
-		Set<ParameterReference> parameters = new HashSet<ParameterReference>();
-		for(SpecificationConstraints sc: specificationConstraints){
-			Set<ParameterReference> ap = sc.getActiveParameters();
-			if(ap != null){
-				parameters.addAll(ap);
-			}
-		}
-		return parameters;
-	}
-	
-	
-
-}
