+++ opendse-encoding/src/main/java/net/sf/opendse/encoding/firm/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Package containing the classes used for the implementation encoding using one firm,
+ * holistic constraint set.
+ */
+package net.sf.opendse.encoding.firm;
