+++ opendse-encoding/src/main/java/net/sf/opendse/encoding/firm/variables/CLRRT.java
@@ -1,0 +1,59 @@
+/*******************************************************************************
+ * Copyright (c) 2015 OpenDSE
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ *******************************************************************************/
+package net.sf.opendse.encoding.firm.variables;
+
+import net.sf.opendse.model.Edge;
+import net.sf.opendse.model.ICommunication;
+import net.sf.opendse.model.Resource;
+import net.sf.opendse.model.Task;
+
+public class CLRRT extends Variable implements CommunicationVariable {
+
+	public CLRRT(Task t, Edge l, Resource r0, Resource r1, int s) {
+		super(t,l,r0,r1,s);
+	}
+
+	public Task getTask() {
+		return get(0);
+	}
+
+	public Edge getLink() {
+		return get(1);
+	}
+
+	public Resource getSource() {
+		return get(2);
+	}
+
+	public Resource getDest() {
+		return get(3);
+	}
+	
+	public Integer getStep(){
+		return get(4);
+	}
+	
+	@Override
+	public ICommunication getCommunication() {
+		return (ICommunication)getTask();
+	}
+}
