+++ opendse-encoding/src/main/java/net/sf/opendse/encoding/firm/variables/CR.java
@@ -19,41 +19,28 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  *******************************************************************************/
-package net.sf.opendse.encoding.old.variables;
+package net.sf.opendse.encoding.firm.variables;
 
-import net.sf.opendse.model.Edge;
 import net.sf.opendse.model.ICommunication;
 import net.sf.opendse.model.Resource;
 import net.sf.opendse.model.Task;
 
-public class CLRRT extends Variable implements CommunicationVariable {
+public class CR extends Variable implements CommunicationVariable {
 
-	public CLRRT(Task t, Edge l, Resource r0, Resource r1, int s) {
-		super(t,l,r0,r1,s);
+	public CR(Task c, Resource r) {
+		super(c,r);
 	}
 
-	public Task getTask() {
+	public Task getC() {
 		return get(0);
 	}
 
-	public Edge getLink() {
+	public Resource getR() {
 		return get(1);
 	}
 
-	public Resource getSource() {
-		return get(2);
-	}
-
-	public Resource getDest() {
-		return get(3);
-	}
-	
-	public Integer getStep(){
-		return get(4);
-	}
-	
 	@Override
 	public ICommunication getCommunication() {
-		return (ICommunication)getTask();
+		return (ICommunication)getC();
 	}
 }
