+++ opendse-encoding/src/main/java/net/sf/opendse/encoding/firm/variables/EAVI.java
@@ -19,71 +19,32 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  *******************************************************************************/
-package net.sf.opendse.encoding.old.variables;
+package net.sf.opendse.encoding.firm.variables;
 
-import java.util.Arrays;
+import net.sf.opendse.model.Element;
 
-public abstract class Variable {
+public class EAVI extends Variable {
 
-	protected Object[] objects;
-	                 
-	public Variable(Object... objects){
-		super();
-		final int n = objects.length;
-		this.objects = new Object[n+1];
-		this.objects[0] = this.getClass();
-		System.arraycopy(objects, 0, this.objects, 1, n);
+	public EAVI(Element e, String attribute, Object value, Integer index){
+		super(e,attribute,value,index);
 	}
 	
-	@SuppressWarnings("unchecked")
-	public <O> O get(int i){
-		return (O)objects[i+1];
+	public Element getE(){
+		return get(0);
 	}
-	                 
-	           
-	/*
-	 * (non-Javadoc)
-	 * 
-	 * @see java.lang.Object#hashCode()
-	 */
-	@Override
-	public int hashCode() {
-		return Arrays.hashCode(objects);
+	
+	public String getA(){
+		return get(1);
 	}
-
-	/*
-	 * (non-Javadoc)
-	 * 
-	 * @see java.lang.Object#equals(java.lang.Object)
-	 */
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		final Variable other = (Variable) obj;
-		return Arrays.equals(this.objects, other.objects);
+	
+	public Object getV(){
+		return get(2);
 	}
-
-	/*
-	 * (non-Javadoc)
-	 * 
-	 * @see java.lang.Object#toString()
-	 */
-	public String toString() {
-		String s = "";
-		s += this.getClass().getSimpleName();
-		s += "[";
-		for(int i=1; i<objects.length; i++){
-			s += objects[i] + ",";
-		}
-		s = s.substring(0, s.length() - 1);
-		s += "]";
-		return s;
+	
+	public Integer getI(){
+		return get(3);
 	}
-
-
+	
+	
+	
 }
