+++ opendse-encoding/src/main/java/net/sf/opendse/encoding/firm/variables/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Package containing the {@link net.sf.opendse.encoding.firm.variables.Variable}s
+ * used for the formulation of the {@link org.opt4j.satdecoding.Constraint}s
+ * encoding valid implementations.
+ */
+package net.sf.opendse.encoding.firm.variables;
