+++ opendse-encoding/src/main/java/net/sf/opendse/encoding/package-info.java
@@ -1,4 +1,11 @@
 /**
- * To be done 
+ * Package containing the encoding interfaces used for the en- and the decoding
+ * of the implementation constraints. The
+ * {@link net.sf.opendse.encoding.ImplementationEncoding} is used to formulate
+ * the {@link org.opt4j.satdecoding.Constraint}s encoding a valid
+ * implementation, while the
+ * {@link net.sf.opendse.encoding.ImplementationInterpreter} is used for the
+ * decoding of a {@link org.opt4j.satdecoding.Model} containing a variable
+ * assignment satisfying the constraints.
  */
 package net.sf.opendse.encoding;
