+++ opendse-encoding/src/main/java/net/sf/opendse/encoding/specificationconstraints/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Package for the classes responsible for the automatic generation of
+ * {@link org.opt4j.satdecoding.Constraint}s based on attributes defined in the
+ * {@link net.sf.opendse.model.Specification}.
+ */
+package net.sf.opendse.encoding.specificationconstraints;
