+++ opendse-optimization/src/test/java/net/sf/opendse/optimization/test/ConnectSpecificationTest.java
@@ -21,8 +21,8 @@
  *******************************************************************************/
 package net.sf.opendse.optimization.test;
 
-import net.sf.opendse.encoding.constraints.ElementList;
-import net.sf.opendse.encoding.constraints.SpecificationConstraints;
+import net.sf.opendse.encoding.specificationconstraints.ElementList;
+import net.sf.opendse.encoding.specificationconstraints.SpecificationConstraints;
 import net.sf.opendse.io.SpecificationReader;
 import net.sf.opendse.io.SpecificationWriter;
 import net.sf.opendse.model.Application;
