+++ opendse-optimization/src/test/java/net/sf/opendse/optimization/test/VariableClassOrderTest.java
@@ -5,7 +5,7 @@ import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 
-import net.sf.opendse.encoding.old.variables.EAVI;
+import net.sf.opendse.encoding.firm.variables.EAVI;
 import net.sf.opendse.model.Dependency;
 import net.sf.opendse.model.Link;
 import net.sf.opendse.model.Mapping;
