+++ opendse-tutorial/src/main/java/net/sf/opendse/tutorial/Part4.java
@@ -1,6 +1,6 @@
 package net.sf.opendse.tutorial;
 
-import net.sf.opendse.encoding.constraints.SpecificationConstraints;
+import net.sf.opendse.encoding.specificationconstraints.SpecificationConstraints;
 import net.sf.opendse.io.SpecificationWriter;
 import net.sf.opendse.model.Application;
 import net.sf.opendse.model.Architecture;
